bacon: add other gesture for torch (2/2)
The FLIP_CAMERA_SCANCODE is not used at all at current state. We replaced
last week torch arrow down gesture with silent/vibrate/sound functionallity.

Some user want still to have torch via gesture. Arrow up is a bit more natural
for it due that it points visually to the torch on the device.

So introduce new v up gesture to toggle the torch.

For now we just use FLIP_CAMERA_SCANCODE which hat the unused keycode
249 due that otherwise we need to use 256 which is out of unsigned char
keycode in kernel. We can change the whole logic on a later point if we
need to trigger more gestures.

For now this is a reliable way.

REMARK: Third party kernel torch will simply just not enable as long the kernel did
not integrate the new keycode 249

Change-Id: I927ea0950c317bdcb7c0d5633e3053e6d01d8af8